[ADQL] Add a mechanism to declare ADQL features and to set which optional ones are supported or not. All ADQLObject instances now have a new function allowing to return a LanguageFeature describing the corresponding ADQL feature.
Only all geometric functions (e.g. POINT, CONTAINS) and the new ADQL-2.1
function `LOWER` are optional. By default, the ADQL-2.1 parser declare them as
supported. But their support can be changed at anytime thanks
ADQLParser.getSupportedFeatures() and ADQLParser.setSupportedFeatures(...).

In the current state, UDF declaration does not work anymore with the ADQL-2.1
parser. This will be fixed in a further commit. Besides, TAP has not yet been
adapted to use/configure the FeatureSet of the parser to use.

Anyway, the Javadoc of all geometric functions has been updated (so that being
conform to their description in ADQL-2.1) as well as the formatting of all
ADQLObjects (e.g. lines fitting on 80 characters, ...).